# docs/openapi.yaml
openapi: 3.0.3
info:
  title: KizunaLink
  version: 1.1.0
  description: |
    KizunaLink is a Lavalink v4-compatible standalone audio sending node for
    Discord bots. It provides a REST API for track loading, session/player
    management, and a WebSocket endpoint for real-time player events.

    ## Authentication
    Every REST and WebSocket request must include the configured password in
    the `Authorization` header (e.g. `Authorization: youshallnotpass`).
    WebSocket connections additionally send `User-Id` and `Client-Name` headers.
  license:
    name: MIT
    url: https://github.com/nikcodex/Kizunalink/blob/main/LICENSE
servers:
  - url: http://localhost:2333
    description: Local development server
...
security:
  - passwordAuth: []
paths:
...
  /v4/sessions/{sessionId}:
    get:
      tags: [Players]
      summary: Get a session
      parameters:
        - $ref: "#/components/parameters/sessionId"
      responses:
        "200":
          description: Session details
        "404":
          description: Session not found
    patch:
      tags: [Players]
      summary: Update session (resuming / timeout)
      parameters:
        - $ref: "#/components/parameters/sessionId"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                resuming:
                  type: boolean
                timeout:
                  type: integer
      responses:
        "200":
          description: Updated session info
        "404":
          description: Session not found
...
  /v4/sessions/{sessionId}/players/{guildId}:
    parameters:
      - $ref: "#/components/parameters/sessionId"
      - $ref: "#/components/parameters/guildId"
    get:
      tags: [Players]
      summary: Get a player (guild playback state)
      responses:
        "200":
          description: Player state
          content:
            application/json:
              schema:
                type: object
        "404":
          description: Player or session not found
    patch:
      tags: [Players]
      summary: Update a player (track, filters, voice, volume, position)
      parameters:
        - name: noReplace
          in: query
          schema:
            type: boolean
          description: If true, do not replace a currently playing track
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlayerUpdate"
      responses:
        "200":
          description: Updated player state
          content:
            application/json:
              schema:
                type: object
        "400":
          description: Invalid request (disabled filters, partial voice state)
        "404":
          description: Session not found
    delete:
      tags: [Players]
      summary: Destroy a player
      responses:
        "204":
          description: Player destroyed
        "404":
          description: Session not found
...
components:
  securitySchemes:
    passwordAuth:
      type: apiKey
      in: header
      name: Authorization
      description: The configured server password (sent as a plain header value)
  parameters:
    sessionId:
      name: sessionId
      in: path
      required: true
      schema:
        type: string
      description: The Lavalink session ID
    guildId:
      name: guildId
      in: path
      required: true
      schema:
        type: string
      description: The Discord guild (server) ID
  schemas:
    PlayerUpdate:
      type: object
      description: Partial player update — all fields optional.
      properties:
        track:
          $ref: "#/components/schemas/Track"
          description: Encoded track to play
        encodedTrack:
          type: string
          description: Base64-encoded track (or null/clear)
        identifier:
          type: string
          description: Track URL or search query to load
        position:
          type: integer
          description: Seek position in milliseconds
        endTime:
          type: integer
          nullable: true
          description: Track end time in milliseconds (or null to clear)
        volume:
          type: integer
          description: Volume (0-1000)
        paused:
          type: boolean
        filters:
          type: object
          description: Audio filter configuration
        voice:
          type: object
          description: Voice state update
          properties:
            token:
              type: string
            endpoint:
              type: string
            sessionId:
              type: string
    Track:
      type: object
      properties:
        encoded:
          type: string
        info:
          type: object
          properties:
            identifier:
              type: string
            title:
              type: string
            author:
              type: string
            length:
              type: integer
            isStream:
              type: boolean
            uri:
              type: string
            artworkUrl:
              type: string
            isrc:
              type: string
            sourceName:
              type: string